fix: SectionController 제거 및 LineController로 메서드 이동
